+++ events/messageUpdate.js
@@ -24,7 +24,7 @@ sql.open('./data/Bastion.sqlite');
 
 module.exports = (oldMessage, newMessage) => {
   if (newMessage.content.includes(newMessage.client.token)) {
-    if (oldMessage.guild) {
+    if (newMessage.deletable) {
       newMessage.delete().catch(e => {
         newMessage.client.log.error(e.stack);
       });
@@ -54,9 +54,11 @@ module.exports = (oldMessage, newMessage) => {
   sql.get(`SELECT filterInvite FROM guildSettings WHERE guildID=${newMessage.guild.id}`).then(guild => {
     if (guild.filterInvite === 'true' && !newMessage.guild.members.get(newMessage.author.id).hasPermission('ADMINISTRATOR')) {
       if (/(https:\/\/)?(www\.)?(discord\.gg|discord\.me|discordapp\.com\/invite\/)\/?([a-z0-9-.]+)?/i.test(newMessage.content)) {
-        newMessage.delete().catch(e => {
-          newMessage.client.log.error(e.stack);
-        });
+        if (newMessage.deletable) {
+          newMessage.delete().catch(e => {
+            newMessage.client.log.error(e.stack);
+          });
+        }
       }
     }
   }).catch(e => {
